refactor(site): extract landing page i18n definitions
